refactor: split combinator file
