+++ parse/combinator.go
@@ -1,0 +1,49 @@
+package parse
+
+import (
+	"errors"
+
+	"github.com/yuniruyuni/lang/ast"
+)
+
+// Select combines some target NonTerminals into single NonTerminal.
+// This new NonTerminal checks if targets match current tokens and
+// returns first maching NonTerminal result.
+// If there is no maching NonTerminal, It returns an "invalid tokens" error.
+func Select(cands ...NonTerminal) NonTerminal {
+	return func(at Pos) (Pos, ast.AST, error) {
+		for _, cand := range cands {
+			nx, parsed, err := cand(at)
+			if err == nil {
+				return nx, parsed, nil
+			}
+		}
+		return at, nil, errors.New("invalid tokens")
+	}
+}
+
+// Merger merges some of ASTs into single AST node.
+type Merger func([]ast.AST) ast.AST
+
+// Concat combines NonTerminals sequence into single NonTerminal.
+// This new NonTerminal checks if sequence match from current tokens and
+// call Merger by matched ASTs then the Merger's result return.
+// If there is a non-matched NonTerminal,
+// It returns the error of the NonTerminal.
+func Concat(m Merger, cands ...NonTerminal) NonTerminal {
+	return func(at Pos) (Pos, ast.AST, error) {
+		asts := make([]ast.AST, 0, len(cands))
+
+		nx := at
+		for _, cand := range cands {
+			var parsed ast.AST
+			var err error
+			nx, parsed, err = cand(nx)
+			if err != nil {
+				return at, nil, err
+			}
+			asts = append(asts, parsed)
+		}
+		return nx, m(asts), nil
+	}
+}
